+++ src/components/ReportsPage/HeatmapPage/Tables/AllScorecardHeatmapTable.tsx
@@ -22,7 +22,7 @@ import { parseEntityRef } from '@backstage/catalog-model';
 import { HeatmapTableHeader } from './HeatmapTableHeader';
 import { HeatmapCell } from '../HeatmapCell';
 import { getFormattedScorecardScores, StringIndexable } from '../HeatmapUtils';
-import { defaultComponentRefContext, entityComponentRef, } from '../../../../utils/ComponentUtils';
+import { defaultComponentRefContext, entityComponentRef } from '../../../../utils/ComponentUtils';
 import { filterNotUndefined } from '../../../../utils/collections';
 
 import { GroupByOption, ScoresByIdentifier } from '../../../../api/types';